Introduce BdvLayer interface
This change is made, to get a cleaner model presenter software pattern.

BdvLayer is a model in the sense of model presenter pattern.
It's possible to listen to a BdvLayer. The layer notifies it's
Listeners when it wants to be repainted.

Because of this, the method Extensible.repaint and Extensible.addLayer
becomes obsolete, and is removed.