SegmenterPanel: fix a bug that garbage collects SegmenterPanel too early
The bug occurred when a garbage collection happened before the the first
segmenter was added to the list of segmenters. With an empty list of
segmenters, the SegmenterPanel would only show the AddSegmenterPanel.
AddSegmeneterPanel has bunch of buttons the allows to add a segmenter
to the list of segmenters. These buttons would call a weak listener, that
got garbage collected to early together with the SegmenterPanel and
therefore would be without effect.

Solution, make the SegmenterPanel a JPanel that way it never gets garbage
collected as long as it is needed.