Add BdvLayerLink: shows a BdvLayer in BDV an updates it automatically.
This change is required because the PredictionLayer (which implements
BdvLayer) was garbage collected to early. BdvLayerLink at the same time
keeps a strong reference to the BdvLayer, ensuring that it will not be
garbage collected to early.

BdvLayerLink also nicely combines all the code required to display a
BdvLayer in BDV. It also updates BDV whenever the BdvLayer indicated a
change.