Improve performance of GPU pixel classification by using fixed cell sizes
The performance of the GPU pixel classification algorithm largely depends
on reusing allocated buffers. Changing the size of the blocks the pixel
classification operates on requires the allocation of new buffers. Which
leads to GPU memory overflow and the freeing of all allocated buffers. This
cases a notable performance gap.

Using fixed cell sizes, even on the border of an image guaranties buffer
reuse and avoids the problem described above.