Implement various features related to closing the labkit window
Added a new argument to LabkitImarisPlugin.imageFromImaris(args):
The boolean "CloseLabkitAfterCalculatingResult" indicates whether
to close Labkit after "Compute result and send it to Imaris" is done.

When the labkit window is closed,
dataset.SetParameter("Labkit", "LabkitWindowClosed", "true")
will be called.

Added a new method LabkitImarisPlugin.closeLabkitWindow(args) to be called
analogously to LabkitImarisPlugin.imageFromImaris(args).
We keep a Map from ApplicationID to Labkit window.
When the closeLabkitWindow method is called, it looks to see whether a
window is registered for the corresponding Application ID, and if so,
that window is closed.
This will not yet kill ongoing computations, e.g., if
"Compute result and send it to Imaris" was called, that will just keep
running in the background.