Make SpareRandomAccessIntType thread safe
Reading SparseRandomAccessIntType while using the brush sometimes
returned wrong values. This is because the TLongIntHashMap that
is used internally is not thread safe. It doesn't support concurrent
read and write operations. First solution is to use a ReadWriteLock
to fix this problem, but that's to slow.

The quick solution is to ignore the result of a read operation to
TLongIntHashMap if it's happening interrupted by a write operation.
The newly implemented ReadRetryWriteLock allows to detect if a read
operation is interrupted by an write operation.

A better solution would be to write a thread save TLongIntHashMap.