LabelsLayer: fix concurrency issue that arose with the new BDV version
LabelsLayer uses a cache to translate the indices, used in the index image,
into colors, that are used to display the labels in big data viewer. This
cache was not thread safe. BDV now seems to use multiple threads to render
individual sources. The cache therefore needs to be made thread safe.

The problem only occurred in the deployed Fiji version, probably happens
only with the newer BDV version, used there. The bug happens when one uses
up to 30 different labels.