Use the LabKit keymap manager in main LabKit window.
This is just a toy example: It *adds* an Actions object, set to
the LabKit context, to the existing shortcuts. The actions added
to the Actions are linked to the keymap manager, and the keybindings
can be edited in the preferences dialog.

The keys that the user edits in the preferences dialog are saved
to the labkit config dir ( home / .labkit / keymaps ).

Right now there are just 3 actions:
- show the preferences dialog
- close it
- the dummy example action that prints a useless message.