Configurable shortcut for the planar mode toggle.
To change as little code as possible we simply make an action that
toggles the planar mode button. Then we register an item listener
to this button, so that it is notified when it is toggled
programmatically.

Also, the BasicLabelingComponent has now a 2nd set of actions and
behaviours, if it is provided with a KeymapManager, distinct from
the ones for the BDV.