Manage SciJava Context lifecycle
And fix failing tests on Windows.

This changes all of labkit-ui to eschew usage of the SingletonContext
class in favor of passing the context between components as appropriate,
creating new contexts in relevant situations, and disposing of contexts
when they are finished being used in all cases except for demo code.

It also escapes backslashes of Windows paths hardcoded into macros,
to account for ImageJ macro using backslash as an escape character.

Some of these changes change existing public API, so this commit
increments the major/minor digit of the development version.